refactor: move internal/upstream/ to internal/muxcore/upstream/ (T021b)
