+++ internal/muxcore/owner/owner_serve_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/thebtf/mcp-mux/internal/muxcore/classify"
-	"github.com/thebtf/mcp-mux/internal/upstream"
+	"github.com/thebtf/mcp-mux/internal/muxcore/upstream"
 	"github.com/thejerf/suture/v4"
 )
